fix(86): common screen types refactor in smartphones and tvs
