+++ src/main/java/com/compstore/exception/NotFoundException.java
@@ -5,6 +5,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class NotFoundException extends RuntimeException {
+
     public NotFoundException(String message) {
         super(message);
     }
