+++ src/main/resources/db/changelog/0.1.0/dictionary/004-create-screen-types.yaml
@@ -1,10 +1,10 @@
 databaseChangeLog:
   - changeSet:
-      id: 001-create-smartphone-screen-types
+      id: 004-create-screen-types
       author: grz55
       changes:
         - createTable:
-            tableName: smartphone_screen_types
+            tableName: screen_types
             columns:
               - column:
                   name: id
@@ -14,6 +14,11 @@ databaseChangeLog:
                     primaryKey: true
               - column:
                   name: name
+                  type: varchar(255)
+                  constraints:
+                    nullable: false
+              - column:
+                  name: screen_type_device_type
                   type: varchar(255)
                   constraints:
                     nullable: false
