+++ src/main/resources/db/changelog/0.1.0/dictionary/changelog-0.1.0-dictionary.yaml
@@ -8,3 +8,6 @@ databaseChangeLog:
   - include:
       file: 003-create-brands.yaml
       relativeToChangelogFile: true
+  - include:
+      file: 004-create-screen-types.yaml
+      relativeToChangelogFile: true
