--- src/main/resources/db/changelog/0.1.0/smartphone/002-create-smartphone-operating-systems.yaml
@@ -1,19 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: 002-create-smartphone-operating-systems
-      author: grz55
-      changes:
-        - createTable:
-            tableName: smartphone_operating_systems
-            columns:
-              - column:
-                  name: id
-                  type: uuid
-                  constraints:
-                    nullable: false
-                    primaryKey: true
-              - column:
-                  name: name
-                  type: varchar(255)
-                  constraints:
-                    nullable: false
