+++ src/main/resources/db/changelog/0.1.0/smartphone/changelog-0.1.0-smartphone.yaml
@@ -1,13 +1,10 @@
 databaseChangeLog:
   - include:
-      file: 001-create-smartphone-screen-types.yaml
+      file: 001-create-smartphone-operating-systems.yaml
       relativeToChangelogFile: true
   - include:
-      file: 002-create-smartphone-operating-systems.yaml
+      file: 002-create-smartphones.yaml
       relativeToChangelogFile: true
   - include:
-      file: 003-create-smartphones.yaml
-      relativeToChangelogFile: true
-  - include:
-      file: 004-create-smartphone-relationships.yaml
+      file: 003-create-smartphone-relationships.yaml
       relativeToChangelogFile: true
