+++ src/main/resources/db/changelog/0.1.0/test_data/dictionary/004-insert-data-screen-types.yaml
@@ -1,0 +1,10 @@
+databaseChangeLog:
+  - changeSet:
+      id: 004-insert-data-screen-types
+      author: grz55
+      changes:
+        - loadUpdateData:
+            file: classpath:/db/changelog/0.1.0/test_data/dictionary/screen-types.csv
+            separator: ','
+            tableName: screen_types
+            primaryKey: id
