+++ src/main/resources/db/changelog/0.1.0/test_data/dictionary/changelog-0.1.0-dictionary-data.yaml
@@ -8,3 +8,6 @@ databaseChangeLog:
   - include:
       file: 003-insert-data-brands.yaml
       relativeToChangelogFile: true
+  - include:
+      file: 004-insert-data-screen-types.yaml
+      relativeToChangelogFile: true
