+++ src/main/resources/db/changelog/0.1.0/test_data/dictionary/screen-types.csv
@@ -1,0 +1,11 @@
+id,name,screen_type_device_type
+b9139d62-2ac1-445c-8108-177c2a6c0c6e,AMOLED,SMARTPHONE
+155b2559-8601-4ead-88cd-3f353964c7f1,OLED,SMARTPHONE
+447612dd-e116-4831-aace-497447b23816,IPS,SMARTPHONE
+aabb81d4-5414-4068-8a51-5734248a694b,POLED,SMARTPHONE
+5ccc86c1-5023-46f5-9f7a-683d6f26ff4f,QD-OLED,TV
+58f4d9ba-3b07-44b4-b04c-4743f6f7be89,OLED,TV
+8b4a8f84-4a85-4d94-be67-2f16a9c240f5,QLED,TV
+34ead23f-6338-4b61-a680-1e389365a2c4,Mini LED,TV
+0b58cd1f-4883-40fe-99e5-fa5d296c936f,Full Array LED,TV
+84bca610-73af-4c9a-8255-1c83efe2d6da,LED,TV
