+++ src/main/resources/db/changelog/0.1.0/test_data/smartphone/001-insert-data-smartphone-operating-systems.yaml
@@ -1,6 +1,6 @@
 databaseChangeLog:
   - changeSet:
-      id: 002-insert-data-smartphone-operating-systems
+      id: 001-insert-data-smartphone-operating-systems
       author: grz55
       changes:
         - loadUpdateData:
