--- src/main/resources/db/changelog/0.1.0/test_data/smartphone/001-insert-data-smartphone-screen-types.yaml
@@ -1,10 +1,0 @@
-databaseChangeLog:
-  - changeSet:
-      id: 001-insert-data-smartphone-screen-types
-      author: grz55
-      changes:
-        - loadUpdateData:
-            file: classpath:/db/changelog/0.1.0/test_data/smartphone/smartphone-screen-types.csv
-            separator: ','
-            tableName: smartphone_screen_types
-            primaryKey: id
