+++ src/main/resources/db/changelog/0.1.0/test_data/smartphone/002-insert-data-smartphones.yaml
@@ -1,6 +1,6 @@
 databaseChangeLog:
   - changeSet:
-      id: 003-insert-data-smartphones
+      id: 002-insert-data-smartphones
       author: grz55
       changes:
         - loadUpdateData:
