+++ src/main/resources/db/changelog/0.1.0/test_data/smartphone/changelog-0.1.0-smartphone-data.yaml
@@ -1,10 +1,7 @@
 databaseChangeLog:
   - include:
-      file: 001-insert-data-smartphone-screen-types.yaml
+      file: 001-insert-data-smartphone-operating-systems.yaml
       relativeToChangelogFile: true
   - include:
-      file: 002-insert-data-smartphone-operating-systems.yaml
-      relativeToChangelogFile: true
-  - include:
-      file: 003-insert-data-smartphones.yaml
+      file: 002-insert-data-smartphones.yaml
       relativeToChangelogFile: true
