--- src/main/resources/db/changelog/0.1.0/test_data/smartphone/smartphone-screen-types.csv
@@ -1,5 +1,0 @@
-id,name
-b9139d62-2ac1-445c-8108-177c2a6c0c6e,AMOLED
-155b2559-8601-4ead-88cd-3f353964c7f1,OLED
-447612dd-e116-4831-aace-497447b23816,IPS
-aabb81d4-5414-4068-8a51-5734248a694b,POLED
