+++ src/main/resources/db/changelog/0.1.0/test_data/smartphone/smartphones.csv
@@ -1,4 +1,4 @@
-id,brand_uuid,name,color,screen_size,resolution_height,resolution_width,smartphone_screen_type_uuid,battery_capacity,ram_capacity,storage_capacity,front_camera_mpix_resolution,back_camera_mpix_resolution,operating_system_uuid,processor_brand_uuid,processor_name,price
+id,brand_uuid,name,color,screen_size,resolution_height,resolution_width,screen_type_uuid,battery_capacity,ram_capacity,storage_capacity,front_camera_mpix_resolution,back_camera_mpix_resolution,operating_system_uuid,processor_brand_uuid,processor_name,price
 e4b9acf7-2a81-4acd-8da9-63a110479e34,3268ab3b-21d1-48d2-9332-2fc659c41428,realme 11 Pro+ 5G,BEIGE,6.7,2412,1080,b9139d62-2ac1-445c-8108-177c2a6c0c6e,5000,RAM_12_GB,STORAGE_512_GB,32,200,30fd1bb4-ff79-4b01-822a-4efaca7b7388,aeefd255-c7a6-4d31-acaf-58f0a3f0fde5,MediaTek Dimensity 7050,569.99
 27c4a82c-6c4b-4060-8f15-7503ea056b0e,09e10d51-4fa0-4435-bdaf-aed6513909d1,Samsung Galaxy S23,BLACK,6.1,2400,1080,b9139d62-2ac1-445c-8108-177c2a6c0c6e,3900,RAM_8_GB,STORAGE_256_GB,12,50,30fd1bb4-ff79-4b01-822a-4efaca7b7388,4c94f5c1-2648-411d-95bd-62bac7d139ea,Qualcomm Snapdragon 8 Gen 2,929.99
 567a7747-c34a-4ea0-964f-a4078508bb05,57599bd5-a588-41e3-b9b0-10a6585f8bae,Apple iPhone 14,BLACK,6.1,2532,1170,155b2559-8601-4ead-88cd-3f353964c7f1,null,null,STORAGE_128_GB,12,12,69172fb7-5b4f-4ad0-9756-8e9981c9932a,39baa0dd-5137-4b31-a01a-2016642103a6,Apple A15 Bionic,849.99
