+++ src/main/resources/db/changelog/0.1.0/test_data/tv/001-insert-data-tvs.yaml
@@ -1,6 +1,6 @@
 databaseChangeLog:
   - changeSet:
-      id: 002-insert-data-tvs
+      id: 001-insert-data-tvs
       author: grz55
       changes:
         - loadUpdateData:
