+++ src/main/resources/db/changelog/0.1.0/test_data/tv/changelog-0.1.0-tv-data.yaml
@@ -1,7 +1,4 @@
 databaseChangeLog:
   - include:
-      file: 001-insert-data-tv-screen-types.yaml
-      relativeToChangelogFile: true
-  - include:
-      file: 002-insert-data-tvs.yaml
+      file: 001-insert-data-tvs.yaml
       relativeToChangelogFile: true
