--- src/main/resources/db/changelog/0.1.0/test_data/tv/tv-screen-types.csv
@@ -1,7 +1,0 @@
-id,name
-5ccc86c1-5023-46f5-9f7a-683d6f26ff4f,QD-OLED
-58f4d9ba-3b07-44b4-b04c-4743f6f7be89,OLED
-8b4a8f84-4a85-4d94-be67-2f16a9c240f5,QLED
-34ead23f-6338-4b61-a680-1e389365a2c4,Mini LED
-0b58cd1f-4883-40fe-99e5-fa5d296c936f,Full Array LED
-84bca610-73af-4c9a-8255-1c83efe2d6da,LED
