+++ src/main/resources/db/changelog/0.1.0/tv/001-create-tvs.yaml
@@ -1,6 +1,6 @@
 databaseChangeLog:
   - changeSet:
-      id: 002-create-tvs
+      id: 001-create-tvs
       author: grz55
       changes:
         - createTable:
