+++ src/main/resources/db/changelog/0.1.0/tv/002-create-tv-relationships.yaml
@@ -1,6 +1,6 @@
 databaseChangeLog:
   - changeSet:
-      id: 003-create-tv-relationships
+      id: 002-create-tv-relationships
       author: grz55
       changes:
         - addForeignKeyConstraint:
@@ -12,6 +12,6 @@ databaseChangeLog:
         - addForeignKeyConstraint:
             baseTableName: tvs
             baseColumnNames: screen_type_uuid
-            referencedTableName: tv_screen_types
+            referencedTableName: screen_types
             referencedColumnNames: id
-            constraintName: fk_tvs_tv_screen_types
+            constraintName: fk_tvs_screen_types
