+++ src/main/resources/db/changelog/0.1.0/tv/changelog-0.1.0-tv.yaml
@@ -1,10 +1,7 @@
 databaseChangeLog:
   - include:
-      file: 001-create-tv-screen-types.yaml
+      file: 001-create-tvs.yaml
       relativeToChangelogFile: true
   - include:
-      file: 002-create-tvs.yaml
-      relativeToChangelogFile: true
-  - include:
-      file: 003-create-tv-relationships.yaml
+      file: 002-create-tv-relationships.yaml
       relativeToChangelogFile: true
